# More MemoryCache perf improvements