--- a/src/libraries/Microsoft.Extensions.Caching.Memory/src/CacheEntry.cs
+++ b/src/libraries/Microsoft.Extensions.Caching.Memory/src/CacheEntry.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
@@ -220,7 +221,8 @@ namespace Microsoft.Extensions.Caching.Memory
             }
         }
 
-        internal bool CheckExpired(DateTimeOffset now)
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal bool CheckExpired(in DateTimeOffset now)
         {
             return _isExpired || CheckForExpiredTime(now) || CheckForExpiredTokens();
         }
@@ -235,27 +237,40 @@ namespace Microsoft.Extensions.Caching.Memory
             DetachTokens();
         }
 
-        private bool CheckForExpiredTime(DateTimeOffset now)
+        private bool CheckForExpiredTime(in DateTimeOffset now)
         {
-            if (_absoluteExpiration.HasValue && _absoluteExpiration.Value <= now)
-            {
-                SetExpired(EvictionReason.Expired);
-                return true;
-            }
+            if (_absoluteExpiration == null && _absoluteExpiration == null)
+                return false;
 
-            if (_slidingExpiration.HasValue
-                && (now - LastAccessed) >= _slidingExpiration)
-            {
-                SetExpired(EvictionReason.Expired);
-                return true;
-            }
+            return SlowPath(now);
 
-            return false;
+            bool SlowPath(in DateTimeOffset offset)
+            {
+                if (_absoluteExpiration.HasValue && _absoluteExpiration.Value <= offset)
+                {
+                    SetExpired(EvictionReason.Expired);
+                    return true;
+                }
+
+                if (_slidingExpiration.HasValue
+                    && (offset - LastAccessed) >= _slidingExpiration)
+                {
+                    SetExpired(EvictionReason.Expired);
+                    return true;
+                }
+
+                return false;
+            }
         }
 
-        internal bool CheckForExpiredTokens()
+        private bool CheckForExpiredTokens()
         {
-            if (_expirationTokens != null)
+            if (_expirationTokens == null)
+                return false;
+
+            return SlowPath();
+
+            bool SlowPath()
             {
                 for (int i = 0; i < _expirationTokens.Count; i++)
                 {
@@ -266,8 +281,8 @@ namespace Microsoft.Extensions.Caching.Memory
                         return true;
                     }
                 }
+                return false;
             }
-            return false;
         }
 
         internal void AttachTokens()